Drop rollup in favour of esbuild.
This new way of building the functions means that we avoid having to run rollup twice, which results in a slow build.

I have used Typescript's AST to retrieve the runtimeConfig, then to compile the TS down to a CJS I have added esbuild (which is faster!).

I've also tidied up a few bit around how the GCP Stack reads configs, so instead of having a manifest file per function I just read the functions.json which contains the same information.